makepkg.conf: Make temporary RPATH contain valid path to $PREFIX/lib
This fixes building of some python native extensions, including 'zlib'
and 'binascii', which used to be removed from the resulting package,
because Python tries running each native extension to check whether
it work, and removes any misbehaving shared objects. This requires
executables to have proper RPATH by the time of doing that checks
(the checks happen in package(), before we fixup the RPATH).