CPP-46507 allow to pass explicit python program path from the IDE
LLDBFrontend was moved to the nativeDebug plugin, so python path
relative to the LLDBFrontend binary become invalid.

This change allows supplying a python program path from the IDE.

The old logic is still used if the explicit path wasn't supplied.
This happens, e.g., if one runs a lldb executable bundled with the IDE
instead of LLDBFrontend.